refactor(api): align route naming with REST conventions
